Sum and Factorio are done

+++ week-02/day-1/src/Sum/Sum.java
@@ -10,7 +10,7 @@ public class Sum {
     }
     public static int sum(int num){
         int result = 0;
-        for (byte i = 0; i <= num; i++) {
+        for (int i = 0; i <= num; i++) {
             result = result + i;
         }
         return result;
